article-backend/requirements.txt
Django>=1.8<2.0
psycopg2-binary
djangorestframework>=3.7,<3.8
